Update clean and setup scripts

+++ bin/clean.sh
@@ -1,17 +1,29 @@
 #!/bin/bash
 # Seed builder
 # AUTO_GENERATED (Read only)
-# Use $ bin/clean.sh <clean_files>
+# Use $ bin/clean.sh <cleaning_type docker|files>
 
-CLEAN_FILES=false
+RUNNING=$(sudo docker compose ps --services --filter "status=running")
+if [ $RUNNING -z ]; then
+  echo "ERROR: Before executing bin/coverage.sh, start server with bin/start.sh"
+  exit 1
+fi
 
-if [ $# -ge 1 ]; then CLEAN_FILES=$1; fi
+CLEANING_TYPE=""
+if [ $# -ge 1 ]; then CLEANING_TYPE=$1; fi
 
-echo "== Cleaning unused docker resources"
-sudo docker image prune --force
-sudo docker volume prune --force
+if [ "CLEANING_TYPE" = "" ]; then
+  echo "ERROR: Invalid option, use bin/clean.bat <docker|files>"
+  exit 1
+fi
 
-if [ "$CLEAN_FILES" = true ]; then
+if [ "CLEANING_TYPE" = "docker" ]; then
+  echo "== Cleaning unused docker resources"
+  sudo docker image prune --force
+  sudo docker volume prune --force
+fi
+
+if [ "CLEANING_TYPE" = "files" ]; then
   echo "== Cleaning unused files"
   sudo docker compose exec django /bin/sh -c "python ./bin/scripts/clean_files.py"
 fi
